+++ app/src/main/java/com/example/coffeehouse/ui/state_holder/CartViewModel.java
@@ -32,7 +32,7 @@ public class CartViewModel extends AndroidViewModel {
         return cartRepositoryImpl.getCartList();
     }
 
-    public LiveData<Integer> getTotalPrice(){
+    public LiveData<Double> getTotalPrice(){
         return cartRepositoryImpl.getTotalPrice();
     }
